python/requirements.txt
pandas
numpy
sqlalchemy
mysql-connector-python
python-dotenv
tqdm
